Level_1 : ft_itoa.c : cleanup 󰃢

+++ Level_4/ft_itoa/ft_itoa.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 
-static long int	ft_numlen(int nbr);
-static char		*ft_getnumstr(char *str, unsigned int nbr, long int len);
+static int		ft_numlen(int nbr);
+static char		*ft_getnumstr(char *str, unsigned int nbr, int len);
 
 /*	Converts an integer int oa string of ASCII characters.
  *
@@ -20,7 +20,7 @@ static char		*ft_getnumstr(char *str, unsigned int nbr, long int len);
  *	*/
 char	*ft_itoa(int nbr)
 {
-	long int		len;
+	int		len;
 	char			*numstr;
 
 	len = ft_numlen(nbr);
@@ -52,9 +52,9 @@ char	*ft_itoa(int nbr)
  *		Increment 'len'
  *	Return 'len'
  *	*/
-static long int	ft_numlen(int nbr)
+static int	ft_numlen(int nbr)
 {
-	long int	len;
+	int	len;
 
 	len = 0;
 	if (nbr < 0)
@@ -76,7 +76,7 @@ static long int	ft_numlen(int nbr)
  *		Divide 'nbr' by 10
  *	Return 'numstr'
  *	*/
-static char		*ft_getnumstr(char *numstr, unsigned int nbr, long int len)
+static char		*ft_getnumstr(char *numstr, unsigned int nbr, int len)
 {
 	while (nbr > 0)
 	{
